Refactor: ChatRoom 서비스 분리 (User,Guest)
