+++ make-decision-helper-back/src/main/java/com/example/make_decision_helper/service/chatroom/UserChatRoomService.java
@@ -1,0 +1,10 @@
+package com.example.make_decision_helper.service.chatroom;
+
+import com.example.make_decision_helper.domain.chatroom.dto.RoomResponse;
+
+import java.util.List;
+
+public interface UserChatRoomService extends ChatRoomService{
+
+    List<RoomResponse> findUserRooms();
+}
